test: move gossip helper method to `TestUtils`
